+++ BitBladeGraphics/include/windows/GraphicsLink.h
@@ -4,7 +4,7 @@
 #ifndef GRAPHICS_LINK_H
 #define GRAPHICS_LINK_H
 
-#include "BladeLinkCommon.h"
+#include "BitBladeCommon.h"
 #include <Windows.h>
 #include <mutex>
 #include <atomic>
@@ -15,7 +15,7 @@ public:
 	GraphicsLink();
 	~GraphicsLink();
 
-	void PackInstruction(char functionCode, const char* data, size_t length);
+	void PackInstruction( char functionCode, const char* data, uint16_t length );
 	const char* GetGraphicsInstructions();
 
 	void SendGraphicsStartupEvent(); // reuses resolve objects received irq (blocking)
@@ -33,7 +33,7 @@ private:
 	char* packedInstructions;	// double buffer for sending graphics update
 	char* outputMessageBuffer;	// given packed instructions when finished
 	char* inputMessageBuffer;	// resolved objects received
-	size_t currentPosition;		// Position tracker for writing to the buffer
+	uint16_t currentPosition;		// Position tracker for writing to the buffer
 
 	// memory mapped files
 	HANDLE hOutputBufferHandle;
@@ -55,8 +55,8 @@ private:
 
 	std::atomic<bool> linkStopSignal;
 
-	HANDLE CreateOrConnectEvent(const char* eventName);
-	void CreateOrOpenMemoryMap(const LPCSTR& test, HANDLE& handleOut, char* bufferOut);
+	HANDLE CreateOrConnectEvent( const char* eventName );
+	void CreateOrOpenMemoryMap( const LPCSTR& test, HANDLE& handleOut, char* bufferOut );
 
 	void irqHandlerOnObjectsTransferFinish();
 	void irqHandlerOnInstructionsReceivedSignal();
